+++ src/components/widget/widget.tsx
@@ -1,5 +1,5 @@
 import React, { type ComponentType } from 'react';
-import { createGlobalStyle, StyleSheetManager, ThemeProvider } from 'styled-components'
+import styled, { StyleSheetManager, ThemeProvider } from 'styled-components'
 import { useSuspenseQuery } from '@tanstack/react-query'
 
 import type { Prettify } from '../../types'
@@ -20,8 +20,7 @@ export type ThemeProps = { theme?: ThemeOptions }
 export type PropsWithI18n<P> = Prettify<P & I18nProps>
 export type PropsWithTheme<P> = Prettify<P & ThemeProps>
 
-const GlobalStyle = createGlobalStyle`
-:root {
+const WidgetContainerThemeVariables = styled(WidgetContainer)`
     --color-primary: ${props => props.theme.primaryColor};
     --color-destructive: ${props => props.theme.dangerColor};
     --color-background: ${props => props.theme.backgroundColor};
@@ -37,7 +36,6 @@ const GlobalStyle = createGlobalStyle`
 
     --border-width: ${props => props.theme.borderWidth};
     --radius: ${props => props.theme.borderRadius};
-}
 `
 
 type PrepareFn<P, U> = (
@@ -92,11 +90,10 @@ export function createWidget<P extends {}, U extends {} = P>({
         return (
             <StyleSheetManager>
                 <ThemeProvider theme={theme}>
-                    <GlobalStyle />
                     <I18nProvider defaultMessages={context.i18n} messages={i18n}>
-                        <WidgetContainer {...widgetAttrs}>
+                        <WidgetContainerThemeVariables {...widgetAttrs} className="r5-widget">
                             <Component {...props as U} />
-                        </WidgetContainer>
+                        </WidgetContainerThemeVariables>
                     </I18nProvider>
                 </ThemeProvider>
             </StyleSheetManager>