+++ CHANGELOG.md
@@ -11,6 +11,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Fix phoneNumberField so phone number is not in the payload instead of always valuating it to null
 
+### Added
+
+- Allow to list and remove trusted devices
+-
 ## [1.32.2] - 2025-02-10
 
 ### Fixed