+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Allow to list and remove trusted devices
+
 ### Fixed
 
 - Fix phoneNumberField so phone number is not in the payload instead of always valuating it to null